@@ -8,12 +8,17 @@ return static function (ECSConfig $ecsConfig) use ($defaultEcsConfigurationSetup
     $defaultSkipList = $defaultEcsConfigurationSetup($ecsConfig, __DIR__);
 
     $ecsConfig->paths([
-        'src',
-        'tests',
-        'ecs.php',
+        __DIR__ . '/src',
+        __DIR__ . '/tests',
+        __DIR__ . '/ecs.php',
     ]);
 
-    $skipList = [];
+    $skipList = [
+        'BrandEmbassyCodingStandard\Sniffs\NamingConvention\CamelCapsFunctionNameSniff.ScopeNotCamelCaps' => [
+            __DIR__ . '/src/EnumType.php',
+            __DIR__ . '/tests/Bridges/MyCLabsEnum/MyCLabsEnumBridgeTest.php',
+        ],
+    ];
 
     $ecsConfig->skip(array_merge($defaultSkipList, $skipList));
 };
