@@ -2,19 +2,24 @@
 
 namespace BrandEmbassy\Doctrine\EnumType;
 
+/**
+ * @template TObject of object
+ */
 interface EnumImplementation
 {
-    public function getBaseEnumClassName(): string;
+    public function isClassSupported(string $className): bool;
+
+
+    public function assertClassIsSupported(string $className): void;
 
 
     /**
-     * @return mixed enum
+     * @template T of TObject
+     * @phpstan-param class-string<T> $enumClassName
+     * @phpstan-return T
      */
-    public function convertDatabaseValueToEnum(string $enumClassName, string $databaseValue);
+    public function convertDatabaseValueToEnum(string $enumClassName, bool|float|int|string|null $databaseValue): mixed;
 
 
-    /**
-     * @param mixed $enum
-     */
-    public function convertEnumToDatabaseValue($enum): string;
+    public function convertEnumToDatabaseValue(mixed $enum): bool|float|int|string|null;
 }
